requirements.txt
# Core dependencies
streamlit>=1.28.0
streamlit-authenticator>=0.2.3
PyYAML>=6.0

# LLM API integrations
google-generativeai>=0.3.0
openai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# File handling
python-multipart>=0.0.6

# Security and encryption
cryptography>=41.0.0

# Database (for logging)
sqlite3

# Additional utilities
pathlib2>=2.3.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0